Update to v2 pricing
