+++ config/coingecko.toml
@@ -1,4 +1,4 @@
-[[COINGECKO_ID_EXPLICIT]]
+[COINGECKO_ID_EXPLICIT]
 aave = "aave"
 ada = "cardano"
 abccvx-gauge = "convex-finance"
