+++ config/price.toml
@@ -8,9 +8,12 @@
 #
 # NOTE: Order matters as it is first match wins
 
+# Token symbols to ignore
 [[rules]]
 symbol = "bchsv"
 action = "ignore"
+
+# Token symbols that need to use coingecko
 
 [[rules]]
 symbol = "ada"
@@ -37,8 +40,14 @@ symbol = "btc"
 action = "coingecko"
 
 [[rules]]
+symbol = "memo"
+action = "coingecko"
+
+[[rules]]
 symbol = "xlm"
 action = "coingecko"
+
+# Token symbols that should be remapped to specific tokens
 
 [[rules]]
 symbol = "avax"
@@ -53,6 +62,12 @@ map_chain_to = "arb"
 map_token_id_to = "0x912CE59144191C1204E64559FE8253a0e49E6548"
 
 [[rules]]
+chain = "eth"
+symbol = "cvxFXS"
+action = "defillama"
+map_token_id_to = "0x3432B6A60D23Ca0dFCa7761B7ab56459D9C964D0" # FXS
+
+[[rules]]
 symbol = "ftm"
 action = "defillama"
 map_chain_to = "ftm"
@@ -64,10 +79,75 @@ action = "defillama"
 map_chain_to = "eth"
 map_token_id_to = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2" # WETH
 
+# Substitutions for tokens that are not on defillama
+
+[[rules]]
+chain = "eth"
+symbol = "afrxeth"
+action = "defillama"
+map_token_id_to = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2" # WETH
+
+[[rules]]
+chain = "eth"
+symbol = "ucvxCRV"
+action = "defillama"
+map_token_id_to = "0x62B9c7356A2Dc64a1969e19C23e4f579F9810Aa7" # cvxCRV
+
+[[rules]]
+chain = "eth"
+symbol = "uCRV"
+action = "defillama"
+map_token_id_to = "0x62B9c7356A2Dc64a1969e19C23e4f579F9810Aa7" # cvxCRV
+
+[[rules]]
+chain = "eth"
+symbol = "xALD"
+action = "defillama"
+map_token_id_to = "0xb26C4B3Ca601136Daf98593feAeff9E0CA702a8D" # ALD
+
+# Set token_id for Coinbase tokens
+
 [[rules]]
 chain = "coinbase"
+symbol = "aave"
 action = "defillama"
 map_chain_to = "eth"
+map_token_id_to = "0x7Fc66500c84A76Ad7e9c93437bFc5Ac33E2DDaE9"
+
+[[rules]]
+chain = "coinbase"
+symbol = "forth"
+action = "defillama"
+map_chain_to = "eth"
+map_token_id_to = "0x77FbA179C79De5B7653F68b5039Af940AdA60ce0"
+
+[[rules]]
+chain = "coinbase"
+symbol = "grt"
+action = "defillama"
+map_chain_to = "eth"
+map_token_id_to = "0xc944E90C64B2c07662A292be6244BDf05Cda44a7"
+
+[[rules]]
+chain = "coinbase"
+symbol = "link"
+action = "defillama"
+map_chain_to = "eth"
+map_token_id_to = "0x514910771AF9Ca656af840dff83E8264EcF986CA"
+
+[[rules]]
+chain = "coinbase"
+symbol = "shib"
+action = "defillama"
+map_chain_to = "eth"
+map_token_id_to = "0x95aD61b0a150d79219dCF64E1E6Cc01f0B64C4cE"
+
+[[rules]]
+chain = "coinbase"
+symbol = "nu"
+action = "defillama"
+map_chain_to = "eth"
+map_token_id_to = "0x4fE83213D56308330EC302a8BD641f1d0113A4Cc"
 
 # This is the default action
 # Do not delete unless you know what you are doing
