+++ reckon/price.py
@@ -1,99 +1,563 @@
-import pandas as pd
+import csv
+import defillama as dl
 import coingecko as cg
-from config import TXNS_OUTPUT, PRICE_OUTPUT, STABLECOINS
+import sys
+import tomllib
+from datetime import datetime
+from config import (
+    PRICE_CONFIG,
+    PRICE_INFERRED_OUTPUT,
+    PRICE_MERGED_OUTPUT,
+    PRICE_MISSING_OUTPUT,
+    PRICE_REQ_OUTPUT,
+    PRICE_OUTPUT,
+    PRICE_WORKSHEET_OUTPUT,
+    STABLECOINS,
+    TXNS_OUTPUT,
+    TXNS_PRICE_REQ_OUTPUT,
+)
 
-def compute_value(
-    date, 
-    txn_type,
-    qty, 
-    name, 
-    purchase_token_cost, 
-    purchase_token, 
-    ):
 
-    if txn_type not in ['buy', 'sell', 'income']:
-        return
+class PriceRule:
 
-    # if not math.isnan(usd_cost):
-    #     return usd_cost
 
-    if type(purchase_token) is str and \
-        purchase_token.lower() in STABLECOINS:
-        # print(f'WARN: Pricing a stablecoin for {qty} {name} <= {purchase_token_cost} {purchase_token}')
-        return purchase_token_cost
+    def __init__(self, entry):
 
-    if type(name) is str and \
-        name.lower() in STABLECOINS:
-        # print(f'WARN: Pricing a stablecoin for {qty} {name} <= {purchase_token_cost} {purchase_token}')
-        return qty
+        self.chain = entry['chain'] if 'chain' in entry else '*'
+        self.symbol = entry['symbol'] if 'symbol' in entry else '*'
+        self.token_id = entry['token_id'] if 'token_id' in entry else '*'
+        self.action = entry['action'] if 'action' in entry else 'defillama'
+        self.params = {k: v for k, v in entry.items() if k not in [
+            "symbol",
+            "action",
+            "chain",
+            "token_id"
+        ]}
 
-    price = cg.get_historical_price(name, date.to_pydatetime())
 
-    if price != None:
-        value = price * qty
-        if type(purchase_token) is str and purchase_token != '':
-            purchase_token_price = value / purchase_token_cost
-            cg.save_historical_price(purchase_token, date.to_pydatetime(), purchase_token_price, "Inferred")
-        return value
+    def __repr__(self):
+        return f"{self.__class__.__name__}({self.chain}, {self.symbol}, {self.token_id}, {self.action}, {self.params})"
 
+
+    def _matches_pattern(self, pattern, value):
+        if pattern is None or pattern == "*":
+            return True
+        return pattern.lower() == value.lower()
+
+
+    def match(self, chain, symbol, token_id):
+        return (self._matches_pattern(self.chain, chain) and
+                self._matches_pattern(self.symbol, symbol) and
+                self._matches_pattern(self.token_id, token_id))
+
+
+    def get_price(self, date, chain, symbol, token_id):
+        return (None, None)
+
+
+    def ignore(self):
+        return self.action == "ignore"
+
+
+class CoinGeckoPriceRule(PriceRule):
+
+
+    def get_price(self, date, chain, symbol, token_id):
+        source = "coingecko"
+        date_obj = datetime.strptime(date, "%Y-%m-%d %H:%M:%S")
+        price = cg.get_historical_price(symbol, date_obj)
+        if price == None or price == '':
+            source += " / price not found"
+        return (price, source)
+
+
+class DefiLlamaPriceRule(PriceRule):
+
+    def get_price(self, date, chain, symbol, token_id):
+        source = "defillama"
+
+        # Look up any params and use them
+        if 'map_chain_to' in self.params:
+            chain = self.params['map_chain_to']
+        if 'map_symbol_to' in self.params:
+            symbol = self.params['map_symbol_to']
+        if 'map_token_id_to' in self.params:
+            token_id = self.params['map_token_id_to'] 
+
+        # Map chain names to defillama API
+        dl_chain = None
+        if chain.lower() == 'eth':
+            dl_chain = 'ethereum'
+        elif chain.lower() == 'ftm':
+            dl_chain = 'fantom'
+        elif chain.lower() == 'arb':
+            dl_chain = 'arbitrum'
+        elif chain.lower() == 'avax':
+            dl_chain = 'avax'
+        else:
+            print(
+                f"WARN: No chain mapping for {date} {chain} {symbol} {token_id}")
+            source += " / no chain mapping"
+
+        # Fix missing token ids
+        if token_id == None:
+            print(
+                f"WARN: No token id found for {chain.lower()}:{symbol.lower()}")
+            source += " / no token_id"
+
+        # Get the price
+        price = dl.get_price(date, dl_chain, symbol, token_id)
+        if price == None or price == '':
+            source += " / price not found"
+
+        return (price, source)
+
+
+class PriceRulesConfig:
+
+    def __init__(self):
+        print('Loading price rules config...')
+        # Read in the price config rules
+        self.rules = []
+        with open(PRICE_CONFIG, "rb") as f:
+            rules_toml = tomllib.load(f)["rules"]
+            for entry in rules_toml:
+                action = entry['action'] if 'action' in entry else 'defillama'
+
+                if action == 'defillama':
+                    rule = DefiLlamaPriceRule(entry)
+                elif action == 'coingecko':
+                    rule = CoinGeckoPriceRule(entry)
+                elif action == 'ignore':
+                    rule = PriceRule(entry)
+                else:
+                    print(f"WARN: Unknown action {action} in price config")
+                    sys.exit(1)
+                print(f"- {rule}")
+                self.rules.append(rule)
+
+
+    def find_rule(self, chain, symbol, token_id):
+        for rule in self.rules:
+            if rule.match(chain, symbol, token_id):
+                return rule
+        return None
+
+
+def reset_prices_file():
+    """
+    Reset the prices file to just the header.
+    """
+
+    print("Resetting prices file...")
+
+    with open(PRICE_REQ_OUTPUT, "w", newline="") as csvfile:
+        writer = csv.writer(csvfile)
+        writer.writerow(["date", "chain", "symbol",
+                        "token_id", "price", "source"])
+
+    with open(PRICE_INFERRED_OUTPUT, "w", newline="") as csvfile:
+        writer = csv.writer(csvfile)
+        writer.writerow(["date", "chain", "symbol",
+                        "token_id", "price", "source"])
+
+
+def get_prices():
+    """
+    Get prices for tokens in TXNS_PRICE_REQ_OUTPUT and save to PRICE_REQ_OUTPUT.
+    """
+
+    print("Getting prices...")
+
+    price_rules_config = PriceRulesConfig()
+
+    # Read in txns from TXNS_PRICE_REQ_OUTPUT
+    price_reqs = []
+    with open(TXNS_PRICE_REQ_OUTPUT, "r") as csvfile:
+        next(csvfile)
+        reader = csv.reader(csvfile)
+        for row in reader:
+            price_reqs.append(row)
+
+    # Iterate through txns and get price for each
+    for req in price_reqs:
+        (
+            date,
+            chain,
+            qty,
+            symbol,
+            token_id,
+            txn_type,
+            purchase_token_cost,
+            purchase_token,
+            purchase_token_id,
+        ) = req
+
+        # Let's try to get the price!
+        price = None
+        source = None
+
+        # Get the action for this token request based on the price rules config
+        rule = price_rules_config.find_rule(chain, symbol, token_id)
+        # print(f"- Match {date} {symbol} {token_id} => {rule}")
+
+        if rule.action == "ignore":
+            print(f"- INFO: Ignoring {date} {chain} {symbol} {token_id}")
+            continue
+
+        (price, source) = rule.get_price(date, chain, symbol, token_id)
+
+        # Save price to file
+        with open(PRICE_REQ_OUTPUT, "a", newline="") as csvfile:
+
+            writer = csv.writer(csvfile)
+            writer.writerow([
+                date,
+                chain,
+                symbol,
+                token_id,
+                price,
+                source
+            ])
+
+        # Infer prices if data is available and qquanitities are significant
+        if price != None and \
+                not purchase_token.lower() in STABLECOINS and \
+                purchase_token != None and \
+                purchase_token != '' and \
+                purchase_token_cost != None and \
+                purchase_token_cost != '' and \
+                float(qty) > 0.1 and \
+                float(purchase_token_cost) > 0.1:
+
+            inferred_price = float(qty) * float(price) / \
+                float(purchase_token_cost)
+            with open(PRICE_INFERRED_OUTPUT, "a", newline="") as csvfile:
+                writer = csv.writer(csvfile)
+                writer.writerow([
+                    date,
+                    chain,
+                    purchase_token,
+                    purchase_token_id,
+                    inferred_price,
+                    "inferred"
+                ])
+
+
+# TODO SHould optimize this to be a global variable rather than reading the
+# file each time.
+def get_price(date, chain, symbol, token_id):
+    """
+    Get price for a single token using the PRICE_REQ_OUTPUT file.
+
+    Note that assumes get_prices() has been run which creates the file based
+    on DefiLlama or Coingecko data.
+    """
+    with open(PRICE_MERGED_OUTPUT, "r") as csvfile:
+        next(csvfile)
+        reader = csv.reader(csvfile)
+        for row in reader:
+            if row[0] == date and row[1] == chain and row[2] == symbol and row[3] == token_id:
+                return row[4]
+    return None
+
+
+def merge_inferred_prices():
+    """
+    Merge inferred prices into requested prices.
+
+    Favor prices that were requested over inferred prices.
+    """
+
+    print("Merging requested and inferred prices...")
+
+    price_reqs = []
+    inferred_prices = []
+    merged_prices = []
+    missing_prices = []
+
+    total_count = 0
+    inferred_count = 0
+    missing_count = 0
+
+    with open(PRICE_REQ_OUTPUT, "r") as csvfile:
+        next(csvfile)
+        reader = csv.reader(csvfile)
+        for row in reader:
+            price_reqs.append(row)
+
+    with open(PRICE_INFERRED_OUTPUT, "r") as csvfile:
+        next(csvfile)
+        reader = csv.reader(csvfile)
+        for row in reader:
+            inferred_prices.append(row)
+
+    for req in price_reqs:
+        (date,
+         chain,
+         symbol,
+         token_id,
+         price,
+         source
+         ) = req
+        total_count += 1
+        if "not found" in source:
+            for row in inferred_prices:
+                (inferred_date,
+                 inferred_chain,
+                 inferred_symbol,
+                 inferred_token_id,
+                 inferred_price,
+                 inferred_source
+                 ) = row
+                if inferred_date == date and \
+                        inferred_chain.lower() == chain.lower() and \
+                        inferred_symbol.lower() == symbol.lower() and \
+                        inferred_token_id.lower() == token_id.lower():
+
+                    inferred_count += 1
+                    merged_prices.append([
+                        date,
+                        chain,
+                        symbol,
+                        token_id,
+                        inferred_price,
+                        inferred_source,
+                    ])
+                    break
+
+            missing_count += 1
+            missing_prices.append([
+                date,
+                chain,
+                symbol,
+                token_id,
+                price,
+                "missing",
+            ])
+
+        else:
+            merged_prices.append([
+                date,
+                chain,
+                symbol,
+                token_id,
+                price,
+                source,
+            ])
+
+    with open(PRICE_MERGED_OUTPUT, "w", newline="") as csvfile:
+        writer = csv.writer(csvfile)
+        writer.writerow(["date", "chain", "symbol",
+                        "token_id", "price", "source"])
+        writer.writerows(merged_prices)
+
+    with open(PRICE_MISSING_OUTPUT, "w", newline="") as csvfile:
+        writer = csv.writer(csvfile)
+        writer.writerow(["date", "chain", "symbol",
+                        "token_id", "price", "source"])
+        writer.writerows(missing_prices)
+
+    print(f"- Total: {total_count}")
+    print(f"- Inferred: {inferred_count}")
+    print(f"- Missing: {missing_count}")
+
+
+def create_priced_txns():
+    """
+    Create PRICE_OUTPUT based on buy, sell, income txns with pricing data.
+    """
+
+    print("Creating priced txns file...")
+    with open(TXNS_OUTPUT, "r", newline="") as csvfile:
+        count_total = 0
+        count_ignored = 0
+        count_stablecoin = 0
+        count_dlcg = 0
+        count_missing = 0
+
+        price_txns = []
+        headers = next(csvfile).strip().split(",")
+        reader = csv.reader(csvfile)
+        for row in reader:
+            count_total += 1
+            date = row[headers.index("date")]
+            txn_type = row[headers.index("txn_type")]
+            qty = row[headers.index("qty")]
+            symbol = row[headers.index("symbol")]
+            token_id = row[headers.index("token_id")]
+            purchase_token_cost = row[headers.index("purchase_token_cost")]
+            purchase_token = row[headers.index("purchase_token")]
+            purchase_token_id = row[headers.index("purchase_token_id")]
+            chain = row[headers.index("chain")]
+            project = row[headers.index("project")]
+            txn_name = row[headers.index("txn_name")]
+            wallet = row[headers.index("wallet")]
+            id = row[headers.index("id")]
+
+            # Skip non-buy, sell, income txns
+            if txn_type.lower() not in ['buy', 'sell', 'income']:
+                count_ignored += 1
+                continue
+
+            # Handle purchased in stablecoins
+            if purchase_token.lower() in STABLECOINS:
+                usd_value = purchase_token_cost
+                price_txns.append([
+                    date,
+                    txn_type,
+                    qty,
+                    symbol,
+                    token_id,
+                    usd_value,
+                    purchase_token_cost,
+                    purchase_token,
+                    purchase_token_id,
+                    chain,
+                    project,
+                    txn_name,
+                    wallet,
+                    id,
+                    'stablecoin/purchase',
+                ])
+                count_stablecoin += 1
+                continue
+
+            # Handle sold or income in stablecoins
+            if symbol.lower() in STABLECOINS:
+                usd_value = qty
+                price_txns.append([
+                    date,
+                    txn_type,
+                    qty,
+                    symbol,
+                    token_id,
+                    usd_value,
+                    purchase_token_cost,
+                    purchase_token,
+                    purchase_token_id,
+                    chain,
+                    project,
+                    txn_name,
+                    wallet,
+                    id,
+                    'stablecoin/sellincome',
+                ])
+                count_stablecoin += 1
+                continue
+
+            price = get_price(date, chain, symbol, token_id)
+
+            if price != None and price != '':
+                if qty != None and qty != '':
+                    usd_value = float(qty) * float(price)
+                else:
+                    print(
+                        f"WARN: Missing qty {date} {chain} {symbol} {token_id}")
+                    usd_value = '0.00'
+                price_txns.append([
+                    date,
+                    txn_type,
+                    qty,
+                    symbol,
+                    token_id,
+                    usd_value,
+                    purchase_token_cost,
+                    purchase_token,
+                    purchase_token_id,
+                    chain,
+                    project,
+                    txn_name,
+                    wallet,
+                    id,
+                    'dlcg',
+                ])
+                count_dlcg += 1
+            else:
+                price_txns.append([
+                    date,
+                    txn_type,
+                    qty,
+                    symbol,
+                    token_id,
+                    'MISSING',
+                    purchase_token_cost,
+                    purchase_token,
+                    purchase_token_id,
+                    chain,
+                    project,
+                    txn_name,
+                    wallet,
+                    id,
+                    'MISSING',
+                ])
+                count_missing += 1
+
+        print(f"- Total txns: {count_total}")
+        print(f"- Ignored txns: {count_ignored}")
+        print(f"- Stablecoin txns: {count_stablecoin}")
+        print(f"- DLCG txns: {count_dlcg}")
+        print(f"- Missing txns: {count_missing}")
+
+        with open(PRICE_OUTPUT, "w", newline="") as csvfile:
+            writer = csv.writer(csvfile)
+            writer.writerow([
+                "date",
+                "txn_type",
+                "qty",
+                "symbol",
+                "token_id",
+                "usd_value",
+                "purchase_token_cost",
+                "purchase_token",
+                "purchase_token_id",
+                "chain",
+                "project",
+                "txn_name",
+                "wallet",
+                "id",
+                "source",
+            ])
+            writer.writerows(price_txns)
+
+
+def create_worksheet():
+    """
+    Create worksheet for manual price entry.
+    """
+
+    with open(PRICE_OUTPUT, newline="", mode="r") as csvfile:
+        reader = csv.reader(csvfile)
+        headers = next(reader)
+        missing = [row for row in reader if row[int(headers.index("source"))] == "MISSING"]
+
+    with open(PRICE_WORKSHEET_OUTPUT, newline="", mode="w") as csvfile:
+        writer = csv.writer(csvfile)
+        writer.writerow([
+            'date',
+            'symbol',
+            'chain',
+            'token_id',
+            'timestamp',
+        ])
+        for row in missing:
+            writer.writerow([
+                row[int(headers.index("date"))],
+                row[int(headers.index("symbol"))],
+                row[int(headers.index("chain"))],
+                row[int(headers.index("token_id"))],
+                dl.get_timestamp_from_date(row[int(headers.index("date"))]),
+            ])
 
 
 def main():
-    """
-    From txns.csv
-    date,txn_type,qty,symbol,purchase_token_cost,purchase_token,txn_name,chain,project,wallet,token_id,purchase_token_id,id,url
-    """
-    df = pd.read_csv(TXNS_OUTPUT)
 
-    df = df.astype({
-        'date': 'datetime64[ns]',
-        'txn_type': 'category',
-        'qty': 'float64',
-        'symbol': 'category',
-        'purchase_token_cost': 'float64',
-        'purchase_token': 'category',
-        'txn_name': 'category',
-        'chain': 'category',
-        'project': 'category',
-        'wallet': 'category',
-        'token_id': 'string',
-        'purchase_token_id': 'string',
-        'id': 'string',
-        'url': 'string',
-    })
-
-    df['usd_value'] = df.apply(
-        lambda x: (compute_value(
-            x['date'], 
-            x['txn_type'],
-            x['qty'], 
-            x['symbol'], 
-            x['purchase_token_cost'],
-            x['purchase_token'],
-        )),
-        axis=1)
-
-    df[[
-        'date',
-        'txn_type',
-        'qty',
-        'symbol',
-        'purchase_token_cost',
-        'purchase_token',
-        'usd_value',
-        'txn_name',
-        'chain',
-        'project',
-        'wallet',
-        'id',
-        'url',
-    ]].to_csv(PRICE_OUTPUT, index=False)
-
-    # TODO Should create an output file of unpriced transactions.
-    #      Especially if it can be used directly to fill price_cache.csv
-
-    cg.clean_caches()
+    reset_prices_file()
+    get_prices()
+    merge_inferred_prices()
+    create_priced_txns()
+    create_worksheet()
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
